eth: split api.go to {admin,debug,miner}_api.go
